% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGenomeVectors.R
\name{AddGenomeVectors}
\alias{AddGenomeVectors}
\title{Generates phyletic vectors of each genome in a group, which stores the
count of appearences of a GO term in that genome.}
\usage{
AddGenomeVectors(defs, anno, genome.names, someGV = NULL)
}
\arguments{
\item{defs}{an enriched KOMODO2-type list object containing at least the
fields named in parameters `anno` and `genome.names`, as well as
a field named `ontology` (containing the name of the ontology to
use). Depending on the ontology other fields may be needed.}

\item{anno}{name of the field in `defs` containing a list with annotation
from the original input data}

\item{genome.names}{name of the field in `defs` containing a char vector with
the names of the genomes to select.}

\item{someGV}{data.frame containing a previously processed set of genome
vectors.}
}
\value{
list with the phyletic vector of each genome of the input group.
}
\description{
Generates phyletic vectors of each genome in a group, which stores the
count of occurrences of a GO term in that genome.
}
