% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data for the KOMODO2 workflow}
\usage{
clean_data(defs)
}
\arguments{
\item{defs}{an enriched KOMODO2-type list object (see Details).}
}
\value{
updated \code{defs} list containing information from parsed
genome maps (e.g., for test and back genomes if `type == "significance"`)
}
\description{
This script implements the second step of the LCFD workflow of KOMODO2.
It is responsible for dealing with data inconsistencies, including
missing values, outliers and undesired characteres, as well as data
merging. It also preprocesses data to allow for more flexible inputs form
the user, such as automatically converting common annotation output to a
single standard format.
}
\details{
The script expects enriched `KOMODO2`-type lists, which are generated by
[load_data()].
}
