% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_analysis.R
\name{do_analysis}
\alias{do_analysis}
\title{Perform analysis for the KOMODO2 workflow}
\usage{
do_analysis(defs)
}
\arguments{
\item{defs}{an enriched KOMODO2-type list object (see Details).}
}
\description{
This script implements the third step of the LCFD workflow of KOMODO2.
It is responsible for performing the actual analysis and generating
all the processed data and tables.
}
\details{
The script expects enriched `KOMODO2`-type lists generated after running
[load_data()] -> [clean_data()].
}
