% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_and_update_packages.R
\name{install_and_update_packages}
\alias{install_and_update_packages}
\title{Install and update dependencies from Suggests field}
\usage{
install_and_update_packages(bioc.args = list())
}
\arguments{
\item{bioc.args}{list containing further arguments to be passed
down to `BiocManager::install()`.}
}
\description{
This function installs the latest versions of all Bioconductor packages
listed in the `Suggests` fields of CALANGO. It is **essential** for
installing the Bioconductor dependencies.
It uses `BiocManager::install()` for installing Bioconductor
packages. Further arguments to this function are passed as a list.
}
\examples{
\dontrun{
  install_and_update_packages()
}
}
