% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_bioc_dependencies.R
\name{install_bioc_dependencies}
\alias{install_bioc_dependencies}
\title{Install Bioconductor dependencies}
\usage{
install_bioc_dependencies(bioc.args = list())
}
\arguments{
\item{bioc.args}{list containing further arguments to be passed
down to `BiocManager::install()`.}
}
\value{
No return value, called for side effects.
}
\description{
This function installs the latest versions of all Bioconductor packages
(listed in the `Suggests` fields of DESCRIPTION). It is **essential** that 
the Bioconductor dependencies be installed for CALANGO to work properly.
It uses `BiocManager::install()` for installing Bioconductor
packages. Further arguments to this function are passed as a list.
}
\examples{
\dontrun{
  install_bioc_dependencies()
}

}
