% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOMODO2_brewer_pal.R
\name{KOMODO2_brewer_pal}
\alias{KOMODO2_brewer_pal}
\title{Adapted from RColorBrewer::brewer.pal()}
\usage{
KOMODO2_brewer_pal(n, name)
}
\arguments{
\item{n}{number of colours to return}

\item{name}{name of palette to use}
}
\description{
Adapted from RColorBrewer::brewer.pal()
}
