% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load and verify all required data for the KOMODO2 workflow}
\usage{
load_data(defs, cores = NULL)
}
\arguments{
\item{defs}{either a KOMODO2-type list object (see Details) or
a path to a text file containing the required definitions.}

\item{cores}{positive integer, how many CPU cores to use (multicore
acceleration does not work in Windows systems). Notice that setting
this parameter will override any `type` field from `defs`.}
}
\value{
updated \code{defs} list containing the information loaded
from the files.
}
\description{
This script represents the first step of the LCFD workflow of KOMODO2.
It separates the data loading, which can be the longest step of a workflow,
from the analysis itself, which is faster and can be redone multiple times.
}
\details{
The script expects a `KOMODO2`-type list, which is a list object containing
at least the following fields:
\itemize{
   \item \code{test.path} (char string): path to the folder containing
   annotation files of the test group
   \item \code{back.path} (char string): path to the folder containing
   annotation files of the background group
   \item \code{x.path} (char string): path to the file containing
   the genomes' attributes (for correlation test)
   \item \code{y.path} (char string): path to the folder containing
   the the genomes and their annotations (for correlation test)
   \item \code{ontology} (char string): which ontology to use. Currently
   accepts "GO" or "Gene Ontology", "KEGG" and "other".
   \item \code{dict.path} (char string): file with the dictionary (terms and
   their meaning) of the ontology, if `ontology` is set as "other".
   \item \code{type} (char string): comparison module to use. Accepts
   "significance" (compares two groups of genomes within an ontology) or
   "correlation" (establishes how much avariable explains the variations
   seen in the genomes).
}

The input definitions can also be passed as a file path. If that is the
case the file must be in a `field = value` format. Blank likes and lines
starting with `#` are ignored. Required fields are the same described for the
`KOMODO2` list described above.
}
