% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{make_report}
\alias{make_report}
\title{Prepare and render the HTML5 report}
\usage{
make_report(defs, render.report = TRUE)
}
\arguments{
\item{defs}{an enriched `CALANGO`-type list generated by `run_CALANGO()`}

\item{render.report}{logical: should the HTML5 report be generated (for
internal use only)}
}
\value{
A (possibly updated) `CALANGO`-type list.
}
\description{
This script generates the HTML5 report based on an enriched
`CALANGO`-type list output by `run_CALANGO()`.
}
