% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_significance_report.R
\name{make_significance_report}
\alias{make_significance_report}
\title{Prepare and render a significance report}
\usage{
make_significance_report(defs, ...)
}
\arguments{
\item{defs}{an enriched `KOMODO2`-type list generated by `run_KOMODO2()`}

\item{...}{other parameters (unused for standalone calls, used internally
by other functions in the package)}
}
\description{
This script generates a significance-type report based on an enriched
`KOMODO2`-type list output by `run_KOMODO2()`.
}
