% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_files.R
\name{retrieve_data_files}
\alias{retrieve_data_files}
\title{Retrieve data files from the Github repository}
\usage{
retrieve_data_files(target.dir, unzip = getOption("unzip"))
}
\arguments{
\item{target.dir}{path to the folder where the files will be saved (
accepts relative and absolute paths)}

\item{unzip}{The unzip method to be used. See the documentation of
\code{utils::unzip()} for details.}
}
\value{
No return value, called for side effects.
}
\description{
This script retrieves relevant data files from the CALANGO project
folder. It will copy the data into a folder containing
directories related to dictionary files, Gene Ontology annotation
files, tree files, etc. Note: you may need to edit the file paths in the 
example scripts contained under the `parameters` subfolder of `target.dir`.
}
\details{
If the `target.dir` provided does not exist it is created
(recursively) by the function.
}
\examples{
\dontrun{
  CALANGO::retrieve_data_files(target.dir = "./data")
}

}
