% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_and_update_packages.R
\name{install_and_update_packages}
\alias{install_and_update_packages}
\title{Install and update dependencies}
\usage{
install_and_update_packages(
  which = "all",
  cran.args = list(),
  bioc.args = list()
)
}
\arguments{
\item{which}{which dependencies to install/update? Accepts
"cran", "bioc" or "all".}

\item{cran.args}{list containing further arguments to be passed
down to `install.packages()`.}

\item{bioc.args}{list containing further arguments to be passed
down to `BiocManager::install()`.}
}
\description{
This function installs the latest versions of all packages
listed in the `Imports` and `Suggests` fields of KOMODO2.
It uses the standard `utils::install.packages()` for CRAN
packages, and `BiocManager::install()` for Bioconductor
dependencies. Further arguments to both functions are passed
as lists.
}
\examples{
\dontrun{
  install_and_update_packages(which = "all")
}
}
