% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{make_report}
\alias{make_report}
\title{Prepare and render the HML5 report}
\usage{
make_report(defs, render.report = TRUE)
}
\arguments{
\item{defs}{an enriched `KOMODO2`-type list generated by `run_KOMODO2()`}

\item{render.report}{logical: should the HTML5 report be generated (for
internal use only)}
}
\value{
A (possibly updated) `KOMODO2`-type list.
}
\description{
This script generates the HTML5 report based on an enriched
`KOMODO2`-type list output by `run_KOMODO2()`.
}
