% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_files.R
\name{retrieve_data_files}
\alias{retrieve_data_files}
\title{Retrieve data files from the Github repository}
\usage{
retrieve_data_files(
  target.dir,
  method = "auto",
  unzip = getOption("unzip"),
  url = "https://github.com/fcampelo/KOMODO2"
)
}
\arguments{
\item{target.dir}{path to the folder where the files will be saved (
accepts relative and absolute paths)}

\item{method}{Method to be used for downloading files. Current download
methods are "internal", "wininet" (Windows only) "libcurl", "wget" and
"curl", and there is a value "auto": see ‘Details’ and ‘Note’ in the
documentation of \code{utils::download.file()}.}

\item{unzip}{The unzip method to be used. See the documentation of
\code{utils::unzip()} for details.}

\item{url}{repository URL. Do not change it unless you really know what
you're doing.}
}
\description{
This script retrieves relevant data files from the KOMODO2 project
repository. It will download the data into a folder containing
directories related to dictionary files, Gene Ontology annotation
files, tree files, etc.
}
\details{
If the `target.dir` provided does not exist it is created
(recursively) by the function.
}
\examples{
\dontrun{
  retrieve_data_files(target.dir = "./data_folder")
}
}
